+++ noctavox/src/tui/layout_minimal.rs
@@ -1,4 +1,4 @@
-use crate::ui_state::{Mode, ProgressDisplay, UiState};
+use crate::ui_state::{LibraryView, Mode, Pane, ProgressDisplay, UiState};
 use ratatui::layout::{Constraint, Layout, Rect};
 
 pub struct LayoutMinimal {
@@ -23,7 +23,7 @@ impl LayoutMinimal {
             false => 0,
             true => match (state.get_progress_display(), area.height > 20) {
                 (ProgressDisplay::ProgressBar, _) | (_, false) => 3,
-                _ => (area.height as f32 * 0.15).ceil() as u16,
+                _ => (area.height as f32 * 0.12).ceil() as u16,
             },
         };
 
@@ -34,13 +34,22 @@ impl LayoutMinimal {
         ])
         .areas(area);
 
-        let [_upper_pad, upper_block, widget, _bottom_pad] = Layout::vertical([
+        let block_h = match state.get_pane() {
+            Pane::TrackList => get_block_height(state.get_legal_songs().len(), area),
+            _ => match state.get_sidebar_view() {
+                LibraryView::Albums => get_block_height(state.albums.len(), area),
+                LibraryView::Playlists => get_block_height(state.playlists.len(), area),
+            },
+        } as u16;
+
+        let [_upper_pad, upper_block, _, widget_spacing, _bottom_pad] = Layout::vertical([
             Constraint::Percentage(20),
-            Constraint::Fill(1),
+            Constraint::Length(block_h),
+            Constraint::Min(1),
             Constraint::Length(widget_h),
             Constraint::Percentage(match is_progress_display {
-                true => 15,
-                false => 20,
+                true => 10,
+                false => 15,
             }),
         ])
         .areas(main_area);
@@ -49,6 +58,13 @@ impl LayoutMinimal {
             Layout::vertical([Constraint::Length(search_height), Constraint::Fill(100)])
                 .areas(upper_block);
 
+        let [_, widget, _] = Layout::horizontal([
+            Constraint::Percentage(10),
+            Constraint::Fill(1),
+            Constraint::Percentage(10),
+        ])
+        .areas(widget_spacing);
+
         LayoutMinimal {
             search_bar,
             content: song_window,
@@ -56,3 +72,15 @@ impl LayoutMinimal {
         }
     }
 }
+
+fn get_block_height(len: usize, area: Rect) -> usize {
+    (len + 4).clamp(3, (area.height as f64 * 0.5).ceil() as usize)
+}
+
+// Vertically Centered:
+//
+//   Constraint::Percentage(20),
+//   Constraint::Length(x as u16),
+//   Constraint::Length(1),
+//   Constraint::Length(widget_h),
+//   Constraint::Fill(1),