+++ noctavox/src/tui/widgets/tracklist/generic_tracklist.rs
@@ -1,7 +1,7 @@
 use crate::{
     library::SongInfo,
     tui::widgets::tracklist::{CellFactory, create_standard_table, get_title},
-    ui_state::{Pane, UiState},
+    ui_state::{LayoutStyle, Pane, UiState},
 };
 use ratatui::{
     style::Stylize,
@@ -27,18 +27,27 @@ impl StatefulWidget for GenericView {
             .map(|(idx, song)| {
                 let is_multi_selected = state.get_multi_select_indices().contains(&idx);
 
-                let index = CellFactory::index_cell(&theme, idx, is_multi_selected);
+                let index =
+                    CellFactory::index_cell(&theme, &state.get_layout(), idx, is_multi_selected);
                 let icon = CellFactory::status_cell(song, state, is_multi_selected);
                 let title = CellFactory::title_cell(&theme, song.get_title(), is_multi_selected);
                 let artist = CellFactory::artist_cell(&theme, song, is_multi_selected);
                 let filetype = CellFactory::filetype_cell(&theme, song, is_multi_selected);
-                let duration = CellFactory::duration_cell(&theme, song, is_multi_selected);
+                let duration = CellFactory::duration_cell_clean(&theme, song, is_multi_selected);
 
-                match is_multi_selected {
-                    true => Row::new([index, icon, title, artist, filetype, duration])
-                        .fg(theme.text_selected)
-                        .bg(state.theme_manager.active.accent_inactive),
-                    false => Row::new([index, icon, title, artist, filetype, duration]),
+                match state.get_layout() {
+                    LayoutStyle::Traditional => match is_multi_selected {
+                        true => Row::new([index, icon, title, artist, filetype, duration])
+                            .fg(theme.text_selected)
+                            .bg(state.theme_manager.active.accent_inactive),
+                        false => Row::new([index, icon, title, artist, filetype, duration]),
+                    },
+                    LayoutStyle::Minimal => match is_multi_selected {
+                        true => Row::new([index, icon, title, duration])
+                            .fg(theme.text_selected)
+                            .bg(state.theme_manager.active.accent_inactive),
+                        false => Row::new([index, icon, title, duration]),
+                    },
                 }
             })
             .collect::<Vec<Row>>();