+++ noctavox/src/ui_state/mod.rs
@@ -37,6 +37,7 @@ use crate::{
     },
 };
 
+#[derive(PartialEq)]
 pub enum LayoutStyle {
     Traditional,
     Minimal,